requirements.txt
Django==1.11.5
django-allauth==0.33.0
django-angular==1.1.4
django-compressor==2.2
django-fake-model==0.1.4
django-filter==1.0.4
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.6.4
mongoengine==0.14.3
Pillow==4.3.0
psycopg2==2.7.3.1
pymongo==3.5.1
